feat: Store PowerState in site-explorer exploration report
With this change, we grab the `system`s power state (which was already fetched but discarded) and persist it inside the endpoint exploration report.

The power state is exposed via gRPC APIs as well as in carbide-web explored endpoint tables.

The change is backward compatible in the following way:
If a PowerState had not been previously fetched, the assume PowerState is `On`. The next site-explorer run will update it with the actual state.

See merge request nvmetal/carbide!2918