fix(agent): drop a .error file, not a ..error file
https://jirasw.nvidia.com/browse/FORGE-4375

When the agent attempts to populate an *nvue* config from the template,
we now move it to a `<path>.error` file if it fails (so we can go look
at the bad config to see what happens). ...but accidentally, it's going
to a `<path>..error` file. Drop the extra "."

See merge request nvmetal/carbide!2927