fix: cast BMC IP as string when comparing against the DPU BMC IPs stored in the explored managed host table.
fix: cast BMC IP as string when comparing against the DPU BMC IPs stored in the explored managed host table. The DPU BMC IPs are nested in a JSON structure. We could cast them to an inet in psql, but it might be safer (and provide more observability) to cast the address to a string on the rust side.

See merge request nvmetal/carbide!2932