Dev-env test: Build pxe boot artifacts when testing trunk
The dev-env test was set to only build PXE boot artifacts if there had been changes to specific directories. This is only relevant to running the test in the pre-merge scenario.

Now that we can run the test on-demand against trunk (via 'scheduled pipelines'), we need to adjust the conditional PXE build to occur in this scenario as well.

See merge request nvmetal/carbide!2936