fix(api): special de/serialization for ch:64 MAC address for SF/VFs
https://jirasw.nvidia.com/browse/FORGE-4373

In my MAC address validation/typing journey ( https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/2924), it turns out that `ch:64` is actually the MAC address as reported by Mellanox for SFs and VFs off of a PF. This is unfortnuate, because it fails to validate, and also fails to deserialize \[from\] any existing topology data in the database.

Instead of ruining it for all of the other MAC addresses (i.e. instead of just going back to `Strings`), slap a custom de/serializer on the `mac_address` field to check if it's `ch:64`, and if it is, turn it into `00:00:00:00:00:64`.

Over time, the `Deserializer` will no longer be needed as we continue to write out a valid `mac_address::MacAddress` naturally as part of serialization.

Put `mac_address: "ch:64"` back where it should be in templates/test data, and updated `test_deserialize_dpu_info` to verify deserialization works as expected.

See merge request nvmetal/carbide!2933