Fix crash in hardware health when host has zero dpus
Actual bug: If there is no attached DPU machine ID, we'd crash
on unwrap().

Fix: Instead of crashing on unwrap(), just skip setting the dpu fields
in the health data.

Also, There are a lot of other unwrap()s in this file, the only one we
really need is when unwrapping the Response builder, and change that to
a .expect() for clarity. For the rest:

- Use `if let Some(...)` instead of `.is_some()` and `.unwrap()`
- Use HashMap::Entry::or_insert_with to get a mutable reference to what
  we've just inserted, instead of repeatedly calling get_mut().unwrap()
  on the HashMap. This cleans up the code a lot.

See merge request nvmetal/carbide!2942