test: Fix metrics assert in api-test by splitting metrics and logs.
Spent a while checking that it wasn't a state machine race condition,
which seems unlikely.

That leaves a metrics issue. Ken suggested it could be because metrics
are currently a singleton bunched in with logging. Split that out so
each test gets it's own prometheus registry.

See merge request nvmetal/carbide!2956