feat: Report a `PoweredOff` health alert if system is not powered on
With this change, the BMC monitoring subsystem will emit a `PoweredOff` alert if the systems powerstate is not `On`.

We already had a `PowerSupply` alert before which inspected the status of the redfish power subsystem. However most of our other code just looks at the top level power status which is encoded in the `ComputerSystem`. Therefore we should do the same here too.

See merge request nvmetal/carbide!2959