fix(api): export machine counts per golden bundle PCR index + value
https://jirasw.nvidia.com/browse/FORGE-4418

This adds additional metrics to provide the machine count for a given PCR index + value, allowing us to further identify common PCR index values across an environment. This is implemented in a way that gives us a view across all bundles (not just within a specific bundle, but if we wanted that, we could export a second metric w/ the bundle_id as an attribute).

It should be noted that this isn't ALL measurements from a report, but merely the should-be-stable measurements from a golden measurement bundle; doing ALL measurements would be very high cardinality, which we don't want -- we're just focusing on select measurement bundle values.

However, since this will be used to determine if there are unexpectedly noisy PCR registers, there could be some churn for a given PCR index, which would be a signal to us that something unexpected is going on.

See merge request nvmetal/carbide!2969