feat: Return whether a Machine/Partition/NetworkSegment is longer in a state than SLA
With this change, users of FindMachinesByIds/etc APIs will become aware which Forge resources are longer in their state than defined by the SLA.
It uses the same SLA calculation as the metric that is already emitted (e.g. forge_machines_per_state_above_sla).

The data is delivered as a new field `state_sla` in the gRPC API which contains the SLA for the current state as well as whether the SLA had been breached.

The data is not yet shown in admin-cli and web-ui. Will allow this in follow-ups.
Implements parts of https://jirasw.nvidia.com/browse/FORGE-4440

See merge request nvmetal/carbide!2988