fix: relax the vpc name condition so to account for the version mismatch between cloud and site
The old condition for `vpc.name` does not work when considering that a new version of cloud which includes `vpc.metadata.name` needs to work with old and new code versions of the site. In the old site version, `vpc.metadata` does not exist and only in the newer version of site we have this field. Hence, in the new code for the cloud we need to send both `vpc.name `and `vpc.metadata.name`. 
I changed the condition to make sure both values are equal to each other if they exist.

See merge request nvmetal/carbide!3012