Drop ad-hoc KeyedObjectFilter types in favor of ObjectColumnFilter
This is a follow-on to !2152, which introduced ObjectColumnFilter, which
allows a simple way to specify how to filter queries based on a
particular column, while enforcing that the correct type is passed.

This drops the various KeyedObjectFilter types in favor of
ObjectColumnFilter, and adds various Column traits to constrain what
can be passed.

Also make a few improvements:

- Add type bounds to ColumnInfo::ColumnType so the filter can be passed
  to sqlx without requiring a huge `where` clause on every find_by
  function

- Add a TableType associated type to ColumnInfo so that find_by
  functions can constrain what columns can be passed (ie. so you can't
  pass some other table's column to find_by)

- Add a FilterableQueryBuilder newtype wrapper around
  sqlx::QueryBuilder, so that we can DRY up all the code which builds
  WHERE clauses from the ObjectColumnFilter

See merge request nvmetal/carbide!2966