Require a flag for force delete w/ allocated instances
We have seen a few times where users force-delete machines w/o knowing
there is an instance allocated to them, and force-delete happily does
that for you. This commit adds a flag to the admin cli to essentially
force the user to type in additional things to ensure they know they
are force deleting a machine with an allocated instance.

See merge request nvmetal/carbide!3031