revert fix(api): save some vpc db calls when building FlatInterfaceConfigs
*This reverts commit `6ba0aed00352dc864f5b9fbbda025185df866fc0`.*

We're currently in a situation where we've noticed `vpc_prefixes` are not
being populated on DPUs. It may have been like this in the 3 weeks since this
went out, and no one may have noticed. We're not entirely sure if this is the
smoking gun, but this is low hanging fruit we can revert (and, when
re-implementing it, making sure tests exist for the change, which there may
not be, but I'll put the time into that after this).

See merge request nvmetal/carbide!3058