fix(api): update existing test to make sure vpc prefixes are set
We ran into an issue where `vpc_prefixes` were empty on the DPUs, and no one noticed for \~3 weeks in the dev environments, and it only got noticed when it went to canary.

There was an existing test, but, unfortunately, the test didn't go deep enough to check every part of a `FlatInterfaceConfig`, so it wasn't looking at `vpc_prefixes`. The change that we \*think\* was the problem was reverted ( https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/3058), and, now that it has been, lets drop a unit test in to make sure `vpc_prefixes`:

* Aren't empty.
* Contain the expected values.

And going forward, if we run into this situation again, tests will fail.

See merge request nvmetal/carbide!3059