Add migration for updating the admin network in an env
This commit adds a stored procedure to be manually called in the event
we need to update the admin networks prefix range. It ensures the
segment provided is marked admin and honors the num_reserved
prefixes. It also wont allow a smaller range because it is a simple
script. It will chew through every machine_interface_address address
and machine_interface hostname and update them both to whatever the
new range is. It then updates the prefix itself. Please note this will
trash the running kea server so it will have to be restarted too.

I explicitly did not want to add this as usable code in carbide. I did
not want anyone using it. We should probably consider deleting it too
in the future.

See merge request nvmetal/carbide!3037