fix: fine tune how we disable secure boot on DPUs
The procedure to disable secure boot on a DPU (as documented on page 58-59 here: https://docs.nvidia.com/networking/display/nvidia-bluefield-management-and-initial-provisioning.pdf):

Redfish call to disable secure boot
Reboot DPU
Reboot DPU
The logic in the state machine currently follows this procedure once. However, it does not verify and check the secure boot status. We have ended up in scenarios where we have not succesfully disabled secure boot on the DPU. This results in the BFB load failing on the DPU.

We have observed that we should proactively check the secure boot status after step (3) from above. If secure boot is still not disabled, follow the procedure again

Fixes: FORGE-4539

See merge request nvmetal/carbide!3050