Refactor DPU agent instrumentation
The way we currently do metrics/instrumentation in the dpu-agent codebase is a bit of a mess, so this represents, hopefully, a slight improvement.

First, we split out the metrics used by the network monitor into a separate metrics state object, since it's architecturally different from the rest of the metrics.

Next, we rework the initialization of the OpenTelemetry configuration so that we're only doing it once (using a `LazyLock` singleton), and create helper functions to get the single `Meter` we now have. We also move the configuration and initialization out of the FMDS code and into a dedicated module (`instrumentation::config`).

Finally, we arrive at the thing I was trying to fix in the first place, which is to use an `ObservableGauge` for the two timestamps added in !2793 so that they don't expire shortly after being set.

See merge request nvmetal/carbide!2898