Machine lifecycle test: Increase sleep before SSH to instance
We have had a couple of test failures because we tried to SSH to the instance before it was fully up and available. 

Increasing the sleep from 20 min to 30 min to hopefully avoid this until we can implement a the phone-home feature to the test.

See merge request nvmetal/carbide!3097