chore: Apply 30min SLA for HostInit/WaitingForDiscovery state
The state did not utilize a SLA in the past. With manual discovery, an admin had to powercycle the host manually in order for the Machine to boot scout and exit the state.
Therefore the machine could sit indefinitely in this state, without it being a deficiency in automation.

With automated ingestion, the above no longer applies.
Automation is supposed to get the Machine through the state in a timely manner. Therefore a SLA should apply.

See merge request nvmetal/carbide!3100