MLT: Enable ngc cli debug logging
As discussed in this slack thread, adding ngc debug logging to the machine lifecycle test.
https://nvidia.slack.com/archives/C077KR4CF6J/p1728586234739479?thread_ts=1728584494.863009&cid=C077KR4CF6J

Setting the environment variable NGC_CLI_DEBUG_LOG to point to a file puts the logs there. Repeated runs append to that. We'll capture the file as an artifact of the gitlab job.

See merge request nvmetal/carbide!3103