feat: [Machine Validation] FORGE-4581 Selectively filter test cases
fix: \[Machine Validation\] Add missing changes - Run on-demand machine validation on failed machine state

## Description

Handle the false positive; Sometime admin takes a call to ignore the test cases and re-run only successful test cases. Eg: Admin wants to ignore GPU test cases so that the machine can be used for CPU job scheduling. 

The feature allows to select the testcases and run then during On-Demand validation. This allows admin to select the test cases that are required to 

Specifically QA wants to recover the machines that are in failed state because of Machine Validation(QA purpose only) 

See merge request nvmetal/carbide!3076