feat: TPM EK cert is checked against a CA
This addresses FORGE-3280.

This change introduces four new forge-admin-cli commands to add, remove, show CA certificates and to show EK certificates that were not matched against CA certificates.

Also, a change has been made in `discover_machine` gRPC call, where the supplied TPM EK certificate is now parsed instead of being simply stored.

Two tables have been added - one to store the CA certificates, and another one to store the statuses of EK certificates wrt a CA certificate that has been used to sign it. When either a CA certificate is added (or removed) or an EK certificate is added, a scan is made to try and match an EK certificate against a CA certificate or vice versa.

Additionally, a failure cause `MeasurementsCAValidationFailed` has been added to signal a situation when measured boot failed due to an EK not having been signed by a CA. This resulted in some changes in the `MachineStateHandler`.
Other changes in `MachineStateHandler` are related to general cleanup of the code. For example, `ManagedHostStateSnapshot` already contains `ManagedHostState`, so there was no reason to supply it independently to `the attempt_state_transition method()`. Also, I dared to improve the naming of variables slightly to better reflect the semantics.

See merge request nvmetal/carbide!2482