Fix: Optimize bulk instance allocation from admin cli
When testing instance allocation for large number of mocked hosts, I noticed the time does not scale well with the number of machines. Then, I noticed one reason is the use of vector data structure and the fact that the for loop will iterate over previously assigned machines each time for finding a free machine. This MR changes the data structure used to be a dequeue so that each machine gets seen once.

See merge request nvmetal/carbide!3115