Zero-DPU: Support network configuration of zero-DPU instances (try 4)
This adds support for network configuration of instances that have zero
DPUs.

Instead of recording synthetic network status observations like in
!2957, or creating a new column to store unmanaged interface config like
in !3049, this derives the instance network status from our internal
view of the instance network config, by adding two new fields to
InstanceInterfaceConfig: the network prefix gateway, and the MAC
address. It also avoids needing a single UnmanagedTenant network like in
!3112.

First, this change creates a new NetworkSegmentType, "HostInband",
which indicates that this segment is a tenant network where the
instance interfaces are equivalent to the host's. (ie. no overlay
network.)

At ingestion time, if a host has zero DPUs, its non-DPU interfaces are
saved to a `predicted_machine_interfaces` table for that machine.

Next, at DHCP time, when the physical machine makes a DHCP request from
its non-DPU interface, we now know what network segment to use for the
interface, and we migrate the interface from predicted_machine_interfaces
to machine_interfaces, with the network segment/prefix/etc information
filled in.

Then, when assigning an instance, if the requested network segment is
the same as the managed host's physical interface, we check to verify
that the segment is of type HostInband, which is the only type that
allows this configuration (normally the host DPU should be on an
underlay segment where the instance should be in an overlay.) If so, we
copy the network configuration of the host interface to the
InstanceInterfaceConfig instead of assigning it a new address.

This gives us the information we need to render the instance network
status with the IP, gateway, prefix, and MAC information filled in,
without needing network status observations to come in from a DPU agent.

See merge request nvmetal/carbide!3125