fix: FORGE-3280: ek cert status deleted
ek cert status entry is now deleted on machine's force delete.

The absence of this functionality meant that when a CA cert is deleted and then a machine is force deleted, a CA cert could not be inserted until such time that a machine is re-discovered. The problem was manifested in this call - https://gitlab-master.nvidia.com/nvmetal/carbide/-/blob/trunk/api/src/handlers/tpm_ca.rs#L72.

Closes FORGE-3280

See merge request nvmetal/carbide!3155