chore: Use mocks to generate data for site-explorer tests
This is a no-op change to the way site exploration reports are generated
for the site_explorer test, as groundwork for more work I intend to do
to make it simple to create new machines for unit tests. (I'd like to
write tests that create new machines, and have those machines use the
real site-explorer based path for ingestion instead of the legacy
discovery path.)

This changes the `ManagedHostConfig` type to support multi/zero DPU,
adding a new `DpuConfig` type to be a member of ManagedHostConfig. It
also makes From impls for SiteExplorationReport, such that the reports
are consistent: they agree on MAC addresses/serial numbers/etc. By
default they get new, unique monotonically increasing MAC addresses and
serial numbers, so that tests don't "accidentally" pass due to static
data. All fields are overridable if the tests need predictable data.

All the places in api/tests/site_explorer.rs where we are creating mock
endpoint reports in-line are converted to using ManagedHostConfig.

Once this is in place I plan to do followup work to create new host
factory methods for tests that do the minimum amount of site-explorer
ingestion to create machine interfaces/etc for the hosts, and then we
can work on adding more unit-level tests for a lot of the zero-DPU work
we're doing.

See merge request nvmetal/carbide!3171