feat: Format Metadata/Labels consistently on web UI pages
We had 2 ways of formatting label lists:
- via code which build a comma separated list of labels for the instance and VPC overviews
- via a template filter which generated a newline separated list of labels for the instance details page

With this change, the 2nd approach is used everywhere in order to remove duplicated code for label formatting. A small change in formatting is that the filter omits the `:` separator in case a label does not specify a value. That makes it easier to identify visually if the value is missing or empty.

The code for truncating too long labels/values has been migrated to the filter.

In addition, the change adds displaying of full Metadata to the VPC details page.

See merge request nvmetal/carbide!3188