chore: Rename "Override" HealthReport mode into "Replace" mode
Forge site admins can override the state of the host using two modes:
- Merge: This mode adds additional alerts
- Override: This mode allows to fully override the health of a machine into any condition - including the option to mark it as healthy.

The 2nd mode was so far called an "override override". Since this terminology is fairly confusing, this change renames it to a "replace" mode.

Thereby the 2 new override modes are:
- Merge
- Replace

For backward compatibility, some code and database internals still refer to the other mode as "Override". However from an API and metrics perspective, "Replace" is utilized.

See merge request nvmetal/carbide!3138