Fix race condition where zero-dpu hosts get stuck if DHCP runs before site-explorer
This is a cherry-pick of the fix found while writing !3177. In case
that one takes longer to review/accept, this brings in just the fix.

If a machine DHCP's before site-explorer sees it, it can create a
machine_interface placeholder entry for the new interface, with no
machine_id. When site-explorer sees the corresponding BMC, it needs to
migrate that interface to the explored host, rather than creating a new
interface. (Particularly because site-explorer does not know what
network segment the explored interface belongs to... that piece of
information is only discovered once a host DHCP's.)

Tests for this fix are in !3177.

See merge request nvmetal/carbide!3184