tests: Fluent interface for guiding site-explorer mocks
This adds the MockExploredHost fixture, which has a fluent interface for
driving the lifecycle of a mock host, controlling when DHCP and
discovery run, versus when site-explorer is run, etc. See the newly
created tests for examples of how it is used.

It also adds a simple `new_host()` wrapper, which uses the fixture to
create hosts with the requested number of DPUs, without hacks like
creating two managed hosts and manually moving one DPU over, which is
how we were doing this before.

Also, use the new MockExploredHost to isolate a race condition where
a zero-DPU host DHCP's before site-explorer gets a chance to run,
causing the host to not get ingested; and fix that bug.

See merge request nvmetal/carbide!3177