fix(carbide-pxe): add WaitingForMeasurements to known states
https://jirasw.nvidia.com/browse/FORGE-4771

There are a known list of states that we'll allow `get_pxe_instructions` to occur on, and any others will yell at the client.

One of the states needs to be `WaitingForMeasurements`.

I hadn't really thought about this at the time, but we're running into a situaton where machine certificates are expiring, `scout` can't talk to the API, and machines are sitting in `WaitingForMeasurements`. The machines \[sometimes\] get rebooted, and then when `carbide-pxe` goes to `get_pxe_instructions`, it sees its in `WaitingForMeasurements`, and then refuses to go on, which shouldn't be the case.

So, I'm adding it to the pool of other allowed states.

See merge request nvmetal/carbide!3195