Update to HBN 2.3 and refactor representors
HBN 2.3 renamed the devices from `_sf` to `_if` and changed numbering scheme from dpu0, dpu1 to dpu1, dpu3. When deploying the dpu agent, it failed a lot due to these renames. This commit adds a compat layer between old hbn and new hbn so the agent can still run on old hbn versions, and upgrades the tests to use the new version. This commit tries to remove all cases of the hardcoded strings in the code as well.

See merge request nvmetal/carbide!3135