feat: Store BMC response body in separate field in EndpointExplorationError
With this change, the EndpointExplorationError will no longer mingle the error message and HTTP response body in the same field ("Details"), but keep the response body (if available) in a separate field.

That allows displaying the errors in a nicer fashion. The Web UI in particular will try to unpack JSON response bodies and pretty-print them.

See merge request nvmetal/carbide!3203