feat: Record response body and code for EndpointExplorationError::InvalidDpuRedfishBiosResponse
This change records the response body for InvalidDpuRedfishBiosResponse in a separate field in order to show them in a better format.

The change will also only map a redfish error to InvalidDpuRedfishBiosResponse in case the error was triggered by a 5xx status code (and not e.g. by temporary unavailability of the DPU or due to invalid credentials). That would avoid the BMC reset in case its not the most appropriate mitigation.

See merge request nvmetal/carbide!3213