feat: Store and display last endpoint exploration duration
Exploring a host BMC can potentially require a lot of time. If there are lots of slow BMC endpoints around, it will slow down the time for the overall site explorer run.

To better understand which hosts BMCs are slow, this change stores the exploration times in the database and exposes them via gRPC API and Web UI.

We already collected the exploration times before and emitted them as metrics. The metrics however did not allow to see which machines where slow.

See merge request nvmetal/carbide!3219