Fix hosts stuck in Discovered when machine validation is disabled
nvmetal/carbide!3191 introduced an escape hatch to skip machine
validation if it's disabled by configuration, but this escape hatch was
only running after we waited for a host to fully reboot itself (ie.
report RebootCompleted), at which point it would move the state machine
to Discovered. But Discovered also waits for the machine to reboot
(because it assumes that a prior step had just rebooted it), which would
never happen.

This is because every state change bumps the version timestamp of the
machine's state snapshot, meaning every check for `rebooted()` will
start returning false every time we bump the state, until we observe
another reboot. So waiting for a reboot in MachineValidation and then
moving to Discovered without issuing another reboot, will always fail as
the host will not reboot again.

The fix here is to pass a boolean, skip_reboot_wait, to the Discovered
state, and set it to true if we've observed the host having already
rebooted when in the MachineValidation state.

We should have seen this in unit tests, but unfortunately the
create_host_with_machine_validation fixture was artificially calling an
additional reboot_completed action when we wouldn't expect it from a
real host, which was unsticking the tests. So drop the extra
forge_action_control call in the fixture, since real hosts will not be
running scout again at this point.

See merge request nvmetal/carbide!3221