fix: FORGE-4745 check power state of machine before attempting power on/off
This adds a check to make sure the machine is in the expected state before attempting a power on or off.  Some machines ignore such a request and others return an error.  this will normalize the behavior into ignoring the request, but allow the state machine to move on.

See merge request nvmetal/carbide!3240