fix: Restructure client certificate renewal and retry renewal in case of failures
This change moves most client certificate renewal related code from the big `main_loop.rs` file into a separate `cert_renewal` module.

It also fixes a problem that prevented certificate renewal from being retried in case it failed: In the previous version of this code, the `cert_renewal_time` timestamp was set to 5-7 days in the future irrespective of whether cert renewal succeeded or not. With this change, the certificate renewal will be retried in between 1 and 5 minutes.

See merge request nvmetal/carbide!3239