feat(web-ui): FORGE-4369 Add oauth2 support to carbide-web UI
This MR...

- Adds Azure/Entra auth to carbide-web so we can replace the basic-auth we've had up to now.
- Defaults to basic-auth so it can be merged and deployed before all sites have been updated to supply the necessary config and secrets.
- Has ready-to-use defaults for everything except:
  * `CARBIDE_WEB_OAUTH2_CLIENT_SECRET` - A secret used to talk to MS entra/graph
  * `CARBIDE_WEB_PRIVATE_COOKIEJAR_KEY` - A secret used for encrypting the cookie values used for sessions
  * `CARBIDE_WEB_HOSTNAME` - A hostname specific for each site that's needed for redirects

It uses Axum's PrivateCookieJar to store an encrypted session cookie so we can avoid any DB work to track sessions for now.  If we want more info attached, we can add to the cookie(s) without the user being able to get at the contents.  I think the only drawback is centralized expiration of a session, but rotating `CARBIDE_WEB_PRIVATE_COOKIEJAR_KEY` and restarting/deploying services works as a break-glass.  Users who have to re-authenticate would only see a slight delay as they bounce between carbide-web and microsoft to get a new session cookie.

I used a lot of `const` defaults and grab things via envvar and kept going back and forth about whether to add to the actual carbide config.  I'll leave it to the MR review to decide.

See merge request nvmetal/carbide!3214