feat: Serve static pxe content from an nginx server if configured
There are a corresponding set of changes in the forged repository to start the nginx pod. But the carbide changes can be pushed ahead of the forged changes. If the nginx configuration information (in the form of carbide_static_pxe_url in carbide-pxe-config.toml) is not found, we will continue to server all content through the pxe server.

See merge request nvmetal/carbide!3228