fix(admin-cli): always show MeasurementJournal report_id
https://jirasw.nvidia.com/browse/FORGE-4864

The `report_id` was configured to only show when `--extended` was set, but
as time has gone on, it turns out we (as operators) frequently want to see
the `report_id` associated with a given `MeasurementJournal`, and it's
annoying to have to type `--extended` almost every time just to see it. So,
this improves it!

See merge request nvmetal/carbide!3265