fix: download root certificate from pxe server in x86
A while ago we added a change that downloads the root cert from the
pxe server for DPUs. We didnt need it on the x86 until now because it
was built into the image. Now that we are using the x86 builds in the
dev env, we need to download it as well. This commit mirrors what the
DPU is already doing.

Ref: 72cc206d68806305f6bf830065f5b09d619a8fa1

See merge request nvmetal/carbide!2767