fix: FORGE-4946 Fix default crypto provider in ufmclient
The newest rustls requires you to pass a default CryptoProvider every
time you build a TLS client, or set a process-wide one, or it crashes at
runtime.  This was fixed for the main Forge client, but the ufmclient
also builds a TLS client with rustls and did not set a default crypto
provider, and I guess no tests cover this.

See merge request nvmetal/carbide!3282