chore: FORGE-4966: Move measured-boot ID types to forge-uuid
This is the first of probably several MR's to gradually eliminate
dependencies on the api crate, see FORGE-4965. The client and server
both need the UUID types, but we have a crate for that. Move them there.

Note: In order to eventually prevent the admin-cli from needing to
depend on sqlx, make the sqlx dependency optional in the forge-uuid
crate, and have the various derive, FromRow, etc code conditional upon a
"sqlx" feature.

See merge request nvmetal/carbide!3284