fix: turn the host back on after turning it off as part of DPU provisioning
fix: when configuring the bios settings on a DPU, we sometimes turn the host off as part of error handling. After turning the host off, the DPU's BMC is no longer accessible. So, if we cannot create a redfish client for the DPU after turning the host off, try to turn the host back on.

See merge request nvmetal/carbide!3286