chore: Make sqlx dependency optional in a few packages
This is some groundwork to make a later MR to remove the sqlx dependency
from the admin-cli, a bit smaller. (It still depends on sqlx now because
of its dependency on the api crate.) Generally when we have a common
crate with some type definitions, and we want to derive sqlx things like
FromRow, etc, these should be optional and gated behind a sqlx feature
flag so that the admin-cli (and other things like agent/etc) do not need
to depend on sqlx. On my machine I'm currently halving the binary size
by eliminating this dependency.

See merge request nvmetal/carbide!3289