tests: better power state emulation in machine-a-tron
Instead of always immediately rebooting when we get any kind of power
signal, emulate the behavior of a real host by reporting invalid power
states (don't power off if already powered off, don't force-reset if
off, etc) and properly emulating powered-off behavior.

This is some test-driven development to try and fix a bug where if hosts
aren't powered on and have no DPUs, we never power them on during site
exploration and they're stuck off. (I think this is specific to zero-dpu
hosts, as DPUs seem to be powered on when chassis power is applied and
thus we start getting farther in the state machine before the host is
powered on? I need to investigate further.) I don't have a fix for this
yet, but this change will draw out the bug if you simply change the
default power state to PowerState::Off in MachineStateMachine::new.

See merge request nvmetal/carbide!3274