chore: Eliminate dependency from admin to carbide-api
The admin crate depends on carbide-api, which is not what we want... it
makes the binary huge, and adds in a bunch of unnecessary dependencies
like sqlx, which make no sense for a CLI app. This change drops the size
of forge-admin-cli in release mode from 756MB to 460MB on my system.

Most of the work here is in moving the measured_boot models to its own
crate and sharing them between carbide-api and admin. The measured-boot
crate will conditionally include sqlx adapters if the "sqlx" feature is
enabled, and only includes "ToTable" stuff if the "cli" feature is
enabled, which minimizes unwanted code.

Implementation notes:

- Once the models themselves are moved to another crate, most of the
  remaining code in carbide-api is databse logic, so rename the module
  to "db" instead of "models"

- The db module now just puts the CRUD functions as bare functions in
  the module, ie db::bundle::new_with_txn instead of
  MeasurementBundle::new_with_txn. This avoids needing to use the
  newtype pattern to add methods to external types.

- I made api/src/measured_boot private temporarily to see dead code, and
  found that a lot of the code is only used by unit tests. Move that
  code to individual `test_support` sub-modules which are only compiled
  with cfg(test).

See merge request nvmetal/carbide!3290