feat: Emit a PoweredOff alert in site explorer if host is powered down
With this change, site-explorer will also scan the `power_state` attribute of an explored host. When the state is not `On`, it will emit a health alert that will prevent the host from being allocated.

The same alert is already emitted by the hardware-health service. Since site-explorer is however running more frequently and is non-optional, it makes sense to also emit the alert from here.

See merge request nvmetal/carbide!3309