feat: Format BMC response in redfish console and inject links for odata elements
This allows users of the view the BMC response in pretty printed format.
For each link the response (indicated by `@odata.id`), a hyperlink will be
injected that allows to use the redfish console to explore the linked path.
This allows to click through the redfish tree of any BMC.

See merge request nvmetal/carbide!3332