chore: Make carbide-api lib mostly private
The way we're organizing our tests in carbide-api is a bit backwards.
Tests which need access to arbitrary types within a crate should be
declared as a cfg(test) mod within the crate. Rust calls these "unit"
tests, even if they may be used to test multiple areas of code within
the crate. Contrast this with the "tests" folder, which is supposed to
only contain integration tests: It can only see types are marked as
public within a lib.rs, because integration tests should be testing your
public API surface area.

But in reality, carbide-api should have no public API surface area:
nothing should be depending on this crate, except tests. Our "tests"
folder, which is supposed to only contain integration tests, is using
lots and lots of types within the carbide lib, and relying on them all
being marked public.

But Marking things public prevents rust from detecting dead code: It
assumes that if it's public, something external to this repo may
eventually use it, and keeps it around.

If we want working dead code detection, the fix is to mark almost
everything in lib.rs as non-public, except:

- the run and migrate functions, which are needed by main.rs and
  api-test
- Some types related to constructing a carbide config, which are needed
  by api-test. This should eventually be fixed too: api-test may be
  better off declaring configs as TOML literals rather than getting its
  grubby hands all over carbide internals. (for instance we probably
  *want* to excercise the code that parses a carbide config file...
  constructing one from pure rust is "cheating" for an integration test.

As for what used to be in the "tests" folder within carbide-api, move
that to a "tests" module within api/src, gated by cfg(test), so that it
can have crate-level visibility on types, without needing them to be
public at the crate root.

- Move api/tests to api/src/tests, declare it as cfg(test)
- Ensure imports are all via crate::* instead of carbide::*
- Create explicit exports for types we know api-test will be need
  (mainly for constructing a cfg file), and change api-test to use those
  import paths
- Don't initialize logging on every test, as tests are all run within
  one process now
- Also, fix test logging setup to not be so wonky: We simply need a
  single ctor in tests/mod.rs that will initialize trace logging for
  tests, and to tell integration tests to not re-initialize them when
  the api server is launched. We don't need any mutexes or anything.
- Clean up lib.rs by moving some code out into submodules and
  reexporting it (run(), CarbideError, CarbideResult)
- When code is used only by a test, rust will see it as dead code. Mark
  those as cfg(test). (This ought to be rare, because if code is only
  used by code, it's really "test support code".) We should move these
  to their own cfg(test) sub-modules, but for now, leave them in the
  same place.
- When something in a sqlx-mapped struct is dead code, use
  allow(dead_code) to avoid issues with FromRow not working. It'd be
  nice if sqlx "just worked" with _leading_underscore names, but afaict
  it doesn't.
- Use allow(dead_code) in a few other important parts where things are a
  work-in-progress (like IpAllocationStrategy)
- `pub use` exports that are only used by test code need to go into
  cfg(test)
- Delete any other dead code

See merge request nvmetal/carbide!3306